+++ common/reconciliation/common/blobstoreIterator.go
@@ -29,12 +29,8 @@ import (
 
 	"github.com/uber/cadence/common/blobstore"
 	"github.com/uber/cadence/common/pagination"
+	"github.com/uber/cadence/common/reconciliation/types"
 )
-
-// BlobstoreEntity allows to deserialize and validate different type of executions
-type BlobstoreEntity interface {
-	Validate() error
-}
 
 type (
 	blobstoreIterator struct {
@@ -46,7 +42,7 @@ type (
 func NewBlobstoreIterator(
 	client blobstore.Client,
 	keys Keys,
-	entity BlobstoreEntity,
+	entity types.BlobstoreEntity,
 ) ScanOutputIterator {
 	return &blobstoreIterator{
 		itr: pagination.NewIterator(keys.MinPage, getBlobstoreFetchPageFn(client, keys, entity)),
@@ -70,7 +66,7 @@ func (i *blobstoreIterator) HasNext() bool {
 func getBlobstoreFetchPageFn(
 	client blobstore.Client,
 	keys Keys,
-	out BlobstoreEntity,
+	out types.BlobstoreEntity,
 ) pagination.FetchFn {
 	return func(token pagination.PageToken) (pagination.Page, error) {
 		index := token.(int)
@@ -108,16 +104,16 @@ func getBlobstoreFetchPageFn(
 	}
 }
 
-func deserialize(data []byte, entity BlobstoreEntity) (*ScanOutputEntity, error) {
+func deserialize(data []byte, entity types.BlobstoreEntity) (*ScanOutputEntity, error) {
 	soe := &ScanOutputEntity{
-		Execution: entity,
+		Execution: entity.Clone(),
 	}
 
 	if err := json.Unmarshal(data, soe); err != nil {
 		return nil, err
 	}
 
-	if err := soe.Execution.(BlobstoreEntity).Validate(); err != nil {
+	if err := soe.Execution.(types.BlobstoreEntity).Validate(); err != nil {
 		return nil, err
 	}
 	return soe, nil
