+++ common/reconciliation/common/writerIterator_test.go
@@ -37,11 +37,13 @@ import (
 	"github.com/uber/cadence/common/mocks"
 	"github.com/uber/cadence/common/pagination"
 	"github.com/uber/cadence/common/persistence"
+	"github.com/uber/cadence/common/reconciliation/iterator"
+	"github.com/uber/cadence/common/reconciliation/types"
 	"github.com/uber/cadence/common/service/config"
 )
 
 var (
-	testBranchToken   = []byte{1, 2, 3}
+	validBranchToken  = []byte{89, 11, 0, 10, 0, 0, 0, 12, 116, 101, 115, 116, 45, 116, 114, 101, 101, 45, 105, 100, 11, 0, 20, 0, 0, 0, 14, 116, 101, 115, 116, 45, 98, 114, 97, 110, 99, 104, 45, 105, 100, 0}
 	executionPageSize = 10
 	testShardID       = 1
 )
@@ -61,7 +63,7 @@ func (s *WriterIteratorSuite) SetupTest() {
 
 func (s *WriterIteratorSuite) TestWriterIterator() {
 	pr := persistence.NewPersistenceRetryer(getMockExecutionManager(10, 10), nil, common.CreatePersistenceRetryPolicy())
-	pItr := NewPersistenceIterator(pr, executionPageSize, testShardID, ConcreteExecutionType)
+	iterator := iterator.ConcreteExecution(pr, executionPageSize)
 	uuid := "uuid"
 	extension := Extension("test")
 	outputDir, err := ioutil.TempDir("", "TestWriterIterator")
@@ -74,8 +76,9 @@ func (s *WriterIteratorSuite) TestWriterIterator() {
 	s.NoError(err)
 	blobstoreWriter := NewBlobstoreWriter(uuid, extension, client, 10)
 	var outputs []*ScanOutputEntity
-	for pItr.HasNext() {
-		exec, err := pItr.Next()
+	for iterator.HasNext() {
+		exec, err := iterator.Next()
+
 		s.NoError(err)
 		soe := &ScanOutputEntity{
 			Execution: exec,
@@ -85,17 +88,18 @@ func (s *WriterIteratorSuite) TestWriterIterator() {
 	}
 	s.NoError(blobstoreWriter.Flush())
 	s.Len(outputs, 100)
-	s.False(pItr.HasNext())
-	_, err = pItr.Next()
+	s.False(iterator.HasNext())
+	_, err = iterator.Next()
 	s.Equal(pagination.ErrIteratorFinished, err)
 	flushedKeys := blobstoreWriter.FlushedKeys()
 	s.Equal(uuid, flushedKeys.UUID)
 	s.Equal(0, flushedKeys.MinPage)
 	s.Equal(9, flushedKeys.MaxPage)
 	s.Equal(Extension("test"), flushedKeys.Extension)
-	blobstoreItr := NewBlobstoreIterator(client, *flushedKeys, &ConcreteExecution{})
+	blobstoreItr := NewBlobstoreIterator(client, *flushedKeys, &types.ConcreteExecution{})
 	i := 0
 	s.True(blobstoreItr.HasNext())
+
 	for blobstoreItr.HasNext() {
 		exec, err := blobstoreItr.Next()
 		s.NoError(err)
