+++ common/reconciliation/invariant/openCurrentExecution_test.go
@@ -20,11 +20,13 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package invariants
+package invariant
 
 import (
 	"errors"
 	"testing"
+
+	"github.com/uber/cadence/common/reconciliation/types"
 
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -52,7 +54,7 @@ func (s *OpenCurrentExecutionSuite) SetupTest() {
 
 func (s *OpenCurrentExecutionSuite) TestCheck() {
 	testCases := []struct {
-		execution       *common.ConcreteExecution
+		execution       *types.ConcreteExecution
 		getCurrentResp  *persistence.GetCurrentExecutionResponse
 		getCurrentErr   error
 		getConcreteResp *persistence.GetWorkflowExecutionResponse
@@ -175,5 +177,167 @@ func (s *OpenCurrentExecutionSuite) TestCheck() {
 		execManager.On("GetCurrentExecution", mock.Anything).Return(tc.getCurrentResp, tc.getCurrentErr)
 		o := NewOpenCurrentExecution(persistence.NewPersistenceRetryer(execManager, nil, c2.CreatePersistenceRetryPolicy()))
 		s.Equal(tc.expectedResult, o.Check(tc.execution))
+	}
+}
+
+func (s *UtilSuite) TestExecutionStillOpen() {
+	testCases := []struct {
+		getExecResp *persistence.GetWorkflowExecutionResponse
+		getExecErr  error
+		expectError bool
+		expectOpen  bool
+	}{
+		{
+			getExecResp: nil,
+			getExecErr:  &shared.EntityNotExistsError{},
+			expectError: false,
+			expectOpen:  false,
+		},
+		{
+			getExecResp: nil,
+			getExecErr:  errors.New("got error"),
+			expectError: true,
+			expectOpen:  false,
+		},
+		{
+			getExecResp: &persistence.GetWorkflowExecutionResponse{
+				State: &persistence.WorkflowMutableState{
+					ExecutionInfo: &persistence.WorkflowExecutionInfo{
+						State: persistence.WorkflowStateCompleted,
+					},
+				},
+			},
+			getExecErr:  nil,
+			expectError: false,
+			expectOpen:  false,
+		},
+		{
+			getExecResp: &persistence.GetWorkflowExecutionResponse{
+				State: &persistence.WorkflowMutableState{
+					ExecutionInfo: &persistence.WorkflowExecutionInfo{
+						State: persistence.WorkflowStateCreated,
+					},
+				},
+			},
+			getExecErr:  nil,
+			expectError: false,
+			expectOpen:  true,
+		},
+	}
+
+	for _, tc := range testCases {
+		execManager := &mocks.ExecutionManager{}
+		execManager.On("GetWorkflowExecution", mock.Anything).Return(tc.getExecResp, tc.getExecErr)
+		pr := persistence.NewPersistenceRetryer(execManager, nil, c2.CreatePersistenceRetryPolicy())
+		open, err := ExecutionStillOpen(&types.Execution{}, pr)
+		if tc.expectError {
+			s.Error(err)
+		} else {
+			s.NoError(err)
+		}
+		if tc.expectOpen {
+			s.True(open)
+		} else {
+			s.False(open)
+		}
+	}
+}
+
+type UtilSuite struct {
+	*require.Assertions
+	suite.Suite
+}
+
+func TestUtilSuite(t *testing.T) {
+	suite.Run(t, new(UtilSuite))
+}
+
+func (s *UtilSuite) SetupTest() {
+	s.Assertions = require.New(s.T())
+}
+
+func (s *UtilSuite) TestExecutionStillExists() {
+	testCases := []struct {
+		getExecResp  *persistence.GetWorkflowExecutionResponse
+		getExecErr   error
+		expectError  bool
+		expectExists bool
+	}{
+		{
+			getExecResp:  &persistence.GetWorkflowExecutionResponse{},
+			getExecErr:   nil,
+			expectError:  false,
+			expectExists: true,
+		},
+		{
+			getExecResp:  nil,
+			getExecErr:   &shared.EntityNotExistsError{},
+			expectError:  false,
+			expectExists: false,
+		},
+		{
+			getExecResp:  nil,
+			getExecErr:   errors.New("got error"),
+			expectError:  true,
+			expectExists: false,
+		},
+	}
+
+	for _, tc := range testCases {
+		execManager := &mocks.ExecutionManager{}
+		execManager.On("GetWorkflowExecution", mock.Anything).Return(tc.getExecResp, tc.getExecErr)
+		pr := persistence.NewPersistenceRetryer(execManager, nil, c2.CreatePersistenceRetryPolicy())
+		exists, err := ExecutionStillExists(&types.Execution{}, pr)
+		if tc.expectError {
+			s.Error(err)
+		} else {
+			s.NoError(err)
+		}
+		if tc.expectExists {
+			s.True(exists)
+		} else {
+			s.False(exists)
+		}
+	}
+}
+
+func (s *UtilSuite) TestDeleteExecution() {
+	testCases := []struct {
+		deleteConcreteErr error
+		deleteCurrentErr  error
+		expectedFixResult *common.FixResult
+	}{
+		{
+			deleteConcreteErr: errors.New("error deleting concrete execution"),
+			expectedFixResult: &common.FixResult{
+				FixResultType: common.FixResultTypeFailed,
+				Info:          "failed to delete concrete workflow execution",
+				InfoDetails:   "error deleting concrete execution",
+			},
+		},
+		{
+			deleteCurrentErr: errors.New("error deleting current execution"),
+			expectedFixResult: &common.FixResult{
+				FixResultType: common.FixResultTypeFailed,
+				Info:          "failed to delete current workflow execution",
+				InfoDetails:   "error deleting current execution",
+			},
+		},
+		{
+			expectedFixResult: &common.FixResult{
+				FixResultType: common.FixResultTypeFixed,
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		execManager := &mocks.ExecutionManager{}
+		execManager.On("DeleteWorkflowExecution", mock.Anything).Return(tc.deleteConcreteErr).Once()
+		if tc.deleteConcreteErr == nil {
+			execManager.On("DeleteCurrentWorkflowExecution", mock.Anything).Return(tc.deleteCurrentErr).Once()
+		}
+		pr := persistence.NewPersistenceRetryer(execManager, nil, c2.CreatePersistenceRetryPolicy())
+		result := DeleteExecution(&types.ConcreteExecution{}, pr)
+		s.Equal(tc.expectedFixResult, result)
 	}
 }
