+++ common/xdc/nDCHistoryResender.go
@@ -27,6 +27,8 @@ import (
 	"errors"
 	"time"
 
+	"github.com/uber/cadence/common/reconciliation/invariant"
+
 	"github.com/uber/cadence/.gen/go/admin"
 	"github.com/uber/cadence/.gen/go/history"
 	"github.com/uber/cadence/.gen/go/shared"
@@ -38,6 +40,7 @@ import (
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/persistence"
 	checks "github.com/uber/cadence/common/reconciliation/common"
+	"github.com/uber/cadence/common/reconciliation/types"
 	"github.com/uber/cadence/common/service/dynamicconfig"
 )
 
@@ -76,7 +79,7 @@ type (
 		historyReplicationFn  nDCHistoryReplicationFn
 		serializer            persistence.PayloadSerializer
 		rereplicationTimeout  dynamicconfig.DurationPropertyFnWithDomainIDFilter
-		currentExecutionCheck checks.Invariant
+		currentExecutionCheck invariant.Invariant
 		logger                log.Logger
 	}
 
@@ -93,7 +96,7 @@ func NewNDCHistoryResender(
 	historyReplicationFn nDCHistoryReplicationFn,
 	serializer persistence.PayloadSerializer,
 	rereplicationTimeout dynamicconfig.DurationPropertyFnWithDomainIDFilter,
-	currentExecutionCheck checks.Invariant,
+	currentExecutionCheck invariant.Invariant,
 	logger log.Logger,
 ) *NDCHistoryResenderImpl {
 
@@ -311,8 +314,8 @@ func (n *NDCHistoryResenderImpl) fixCurrentExecution(
 	if n.currentExecutionCheck == nil {
 		return false
 	}
-	execution := &checks.CurrentExecution{
-		Execution: checks.Execution{
+	execution := &types.CurrentExecution{
+		Execution: types.Execution{
 			DomainID:   domainID,
 			WorkflowID: workflowID,
 			State:      persistence.WorkflowStateRunning,
