+++ service/history/timerQueueProcessor.go
@@ -27,13 +27,14 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/uber/cadence/common/reconciliation/invariant"
+
 	h "github.com/uber/cadence/.gen/go/history"
 	"github.com/uber/cadence/client/matching"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
 	"github.com/uber/cadence/common/persistence"
-	checks "github.com/uber/cadence/common/reconciliation/common"
 	"github.com/uber/cadence/common/xdc"
 	"github.com/uber/cadence/service/history/config"
 	"github.com/uber/cadence/service/history/queue"
@@ -70,7 +71,7 @@ func newTimerQueueProcessor(
 	historyService *historyEngineImpl,
 	matchingClient matching.Client,
 	queueTaskProcessor task.Processor,
-	openExecutionCheck checks.Invariant,
+	openExecutionCheck invariant.Invariant,
 	logger log.Logger,
 ) queue.Processor {
