+++ service/history/transferQueueProcessor.go
@@ -27,6 +27,8 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/uber/cadence/common/reconciliation/invariant"
+
 	h "github.com/uber/cadence/.gen/go/history"
 	"github.com/uber/cadence/client/history"
 	"github.com/uber/cadence/client/matching"
@@ -34,7 +36,6 @@ import (
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
 	"github.com/uber/cadence/common/persistence"
-	checks "github.com/uber/cadence/common/reconciliation/common"
 	"github.com/uber/cadence/common/xdc"
 	"github.com/uber/cadence/service/history/config"
 	"github.com/uber/cadence/service/history/queue"
@@ -72,7 +73,7 @@ func newTransferQueueProcessor(
 	matchingClient matching.Client,
 	historyClient history.Client,
 	queueTaskProcessor task.Processor,
-	openExecutionCheck checks.Invariant,
+	openExecutionCheck invariant.Invariant,
 	logger log.Logger,
 ) queue.Processor {
