+++ service/worker/scanner/executions/shard/scanner_test.go
@@ -35,6 +35,7 @@ import (
 
 	"github.com/uber/cadence/common/persistence"
 	"github.com/uber/cadence/common/reconciliation/common"
+	"github.com/uber/cadence/common/reconciliation/types"
 )
 
 type ScannerSuite struct {
@@ -87,12 +88,12 @@ func (s *ScannerSuite) TestScan_Failure_NonFirstError() {
 	mockItr.EXPECT().HasNext().DoAndReturn(func() bool {
 		return iteratorCallNumber < 5
 	}).Times(5)
-	mockItr.EXPECT().Next().DoAndReturn(func() (*common.ConcreteExecution, error) {
+	mockItr.EXPECT().Next().DoAndReturn(func() (*types.ConcreteExecution, error) {
 		defer func() {
 			iteratorCallNumber++
 		}()
 		if iteratorCallNumber < 4 {
-			return &common.ConcreteExecution{}, nil
+			return &types.ConcreteExecution{}, nil
 		}
 		return nil, fmt.Errorf("iterator got error on: %v", iteratorCallNumber)
 	}).Times(5)
@@ -125,7 +126,7 @@ func (s *ScannerSuite) TestScan_Failure_NonFirstError() {
 func (s *ScannerSuite) TestScan_Failure_CorruptedWriterError() {
 	mockItr := pagination.NewMockIterator(s.controller)
 	mockItr.EXPECT().HasNext().Return(true).Times(1)
-	mockItr.EXPECT().Next().Return(&common.ConcreteExecution{}, nil).Times(1)
+	mockItr.EXPECT().Next().Return(&types.ConcreteExecution{}, nil).Times(1)
 	mockInvariantManager := common.NewMockInvariantManager(s.controller)
 	mockInvariantManager.EXPECT().RunChecks(gomock.Any()).Return(common.ManagerCheckResult{
 		CheckResultType: common.CheckResultTypeCorrupted,
@@ -158,7 +159,7 @@ func (s *ScannerSuite) TestScan_Failure_CorruptedWriterError() {
 func (s *ScannerSuite) TestScan_Failure_FailedWriterError() {
 	mockItr := pagination.NewMockIterator(s.controller)
 	mockItr.EXPECT().HasNext().Return(true).Times(1)
-	mockItr.EXPECT().Next().Return(&common.ConcreteExecution{}, nil).Times(1)
+	mockItr.EXPECT().Next().Return(&types.ConcreteExecution{}, nil).Times(1)
 	mockInvariantManager := common.NewMockInvariantManager(s.controller)
 	mockInvariantManager.EXPECT().RunChecks(gomock.Any()).Return(common.ManagerCheckResult{
 		CheckResultType: common.CheckResultTypeFailed,
@@ -251,34 +252,34 @@ func (s *ScannerSuite) TestScan_Success() {
 	mockItr.EXPECT().HasNext().DoAndReturn(func() bool {
 		return iteratorCallNumber < 10
 	}).Times(11)
-	mockItr.EXPECT().Next().DoAndReturn(func() (*common.ConcreteExecution, error) {
+	mockItr.EXPECT().Next().DoAndReturn(func() (*types.ConcreteExecution, error) {
 		defer func() {
 			iteratorCallNumber++
 		}()
 		switch iteratorCallNumber {
 		case 0, 1, 2, 3:
-			return &common.ConcreteExecution{
-				Execution: common.Execution{
+			return &types.ConcreteExecution{
+				Execution: types.Execution{
 					DomainID: "healthy",
 				},
 			}, nil
 		case 4, 5, 6:
-			return &common.ConcreteExecution{
-				Execution: common.Execution{
+			return &types.ConcreteExecution{
+				Execution: types.Execution{
 					DomainID: "history_missing",
 					State:    persistence.WorkflowStateCompleted,
 				},
 			}, nil
 		case 7:
-			return &common.ConcreteExecution{
-				Execution: common.Execution{
+			return &types.ConcreteExecution{
+				Execution: types.Execution{
 					DomainID: "orphan_execution",
 					State:    persistence.WorkflowStateCreated,
 				},
 			}, nil
 		case 8, 9:
-			return &common.ConcreteExecution{
-				Execution: common.Execution{
+			return &types.ConcreteExecution{
+				Execution: types.Execution{
 					DomainID: "failed",
 				},
 			}, nil
@@ -287,15 +288,15 @@ func (s *ScannerSuite) TestScan_Success() {
 		}
 	}).Times(10)
 	mockInvariantManager := common.NewMockInvariantManager(s.controller)
-	mockInvariantManager.EXPECT().RunChecks(&common.ConcreteExecution{
-		Execution: common.Execution{
+	mockInvariantManager.EXPECT().RunChecks(&types.ConcreteExecution{
+		Execution: types.Execution{
 			DomainID: "healthy",
 		},
 	}).Return(common.ManagerCheckResult{
 		CheckResultType: common.CheckResultTypeHealthy,
 	}).Times(4)
-	mockInvariantManager.EXPECT().RunChecks(&common.ConcreteExecution{
-		Execution: common.Execution{
+	mockInvariantManager.EXPECT().RunChecks(&types.ConcreteExecution{
+		Execution: types.Execution{
 			DomainID: "history_missing",
 			State:    persistence.WorkflowStateCompleted,
 		},
@@ -310,8 +311,8 @@ func (s *ScannerSuite) TestScan_Success() {
 			},
 		},
 	}).Times(3)
-	mockInvariantManager.EXPECT().RunChecks(&common.ConcreteExecution{
-		Execution: common.Execution{
+	mockInvariantManager.EXPECT().RunChecks(&types.ConcreteExecution{
+		Execution: types.Execution{
 			DomainID: "orphan_execution",
 			State:    persistence.WorkflowStateCreated,
 		}}).Return(common.ManagerCheckResult{
@@ -329,8 +330,8 @@ func (s *ScannerSuite) TestScan_Success() {
 			},
 		},
 	}).Times(1)
-	mockInvariantManager.EXPECT().RunChecks(&common.ConcreteExecution{
-		Execution: common.Execution{
+	mockInvariantManager.EXPECT().RunChecks(&types.ConcreteExecution{
+		Execution: types.Execution{
 			DomainID: "failed",
 		}}).Return(common.ManagerCheckResult{
 		CheckResultType:          common.CheckResultTypeFailed,
@@ -346,8 +347,8 @@ func (s *ScannerSuite) TestScan_Success() {
 
 	mockCorruptedWriter := common.NewMockExecutionWriter(s.controller)
 	mockCorruptedWriter.EXPECT().Add(common.ScanOutputEntity{
-		Execution: &common.ConcreteExecution{
-			Execution: common.Execution{
+		Execution: &types.ConcreteExecution{
+			Execution: types.Execution{
 				DomainID: "history_missing",
 				State:    persistence.WorkflowStateCompleted,
 			}},
@@ -364,8 +365,8 @@ func (s *ScannerSuite) TestScan_Success() {
 		},
 	}).Times(3)
 	mockCorruptedWriter.EXPECT().Add(common.ScanOutputEntity{
-		Execution: &common.ConcreteExecution{
-			Execution: common.Execution{
+		Execution: &types.ConcreteExecution{
+			Execution: types.Execution{
 				DomainID: "orphan_execution",
 				State:    persistence.WorkflowStateCreated,
 			}},
@@ -387,8 +388,8 @@ func (s *ScannerSuite) TestScan_Success() {
 	}).Times(1)
 	mockFailedWriter := common.NewMockExecutionWriter(s.controller)
 	mockFailedWriter.EXPECT().Add(common.ScanOutputEntity{
-		Execution: &common.ConcreteExecution{
-			Execution: common.Execution{
+		Execution: &types.ConcreteExecution{
+			Execution: types.Execution{
 				DomainID: "failed",
 			}},
 		Result: common.ManagerCheckResult{
